+++ tests/Unit/Form/Abstracts/TemporalTestAbstract.php
@@ -41,6 +41,38 @@ abstract class TemporalTestAbstract extends InputTestAbstract
         $this->getComponent()->model($user)->name('name')->toHtml();
     }
 
+    public function testSetWireWithoutOptionWithModelBinding(): void
+    {
+        $user = $this->createUniqueUser();
+        $user->published_at = $this->faker->dateTime;
+        $html = $this->getComponent()->name('published_at')->model($user)->wire()->toHtml();
+        self::assertStringContainsString(' wire:model="user.published_at"', $html);
+    }
+
+    public function testSetWireWithoutOptionWithCustomModel(): void
+    {
+        $user = $this->createUniqueUser();
+        $user->published_at = $this->faker->dateTime;
+        $html = $this->getComponent()->name('published_at')->model($user)->wire(null, 'test')->toHtml();
+        self::assertStringContainsString(' wire:model="test.published_at"', $html);
+    }
+
+    public function testSetWireWithOptionWithModelBinding(): void
+    {
+        $user = $this->createUniqueUser();
+        $user->published_at = $this->faker->dateTime;
+        $html = $this->getComponent()->name('published_at')->model($user)->wire('defer')->toHtml();
+        self::assertStringContainsString(' wire:model.defer="user.published_at"', $html);
+    }
+
+    public function testSetWireWithOptionWithCustomModel(): void
+    {
+        $user = $this->createUniqueUser();
+        $user->published_at = $this->faker->dateTime;
+        $html = $this->getComponent()->name('published_at')->model($user)->wire('defer', 'test')->toHtml();
+        self::assertStringContainsString(' wire:model.defer="test.published_at"', $html);
+    }
+
     public function testDefaultDisplaySuccess(): void
     {
         config()->set(